requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
astroid==2.9.0
async-timeout==3.0.1
attrs==21.2.0
awsebcli==3.20.3
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
bleach==3.3.0
blessed==1.19.0
botocore==1.23.54
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.1
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
configparser==5.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
Django==3.2.12
django-admin-autocomplete-filter==0.6.1
django-admin-interface==0.18.2
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.8.3
django-admin-sortable2==1.0
django-ckeditor==6.0.0
django-colorfield==0.4.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-facebook==6.0.3
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-json-widget==1.1.1
django-multiselectfield==0.1.12
django-paranoid==1.0
django-parler==2.3
django-parler-rest==2.1
django-prettyjson==0.4.1
django-redis==5.0.0
django-simple-history==3.0.0
django-storages==1.11.1
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
easydict==1.9
environ==1.0
facebook-sdk==3.1.0
filelock==3.4.0
firebase-admin==4.5.3
Flask==1.1.4
Flask-Cors==3.0.10
flask-restplus==0.13.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.30.1
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-cloud-translate==3.1.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
graphene==3.0b8
graphene-django==3.0.0b7
graphene-django-optimizer==0.9.1
graphql-core==3.1.6
graphql-relay==3.1.0
grpcio==1.43.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.0
idna==2.10
importlib-resources==5.4.0
inflection==0.5.1
isodate==0.6.0
isort==5.10.1
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.6.0
logger==1.4
lxml==4.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
oauthlib==3.1.0
packaging==20.9
paramiko==2.8.1
pathspec==0.5.9
Pillow==9.1.0
platformdirs==2.4.0
promise==2.3
proto-plus==1.18.1
protobuf==3.15.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.12.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-gettext==4.0
python-memcached==1.59
python-twitter==3.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
reportlab==3.5.67
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
Rx==1.6.1
s3transfer==0.4.2
schedule==1.1.0
semantic-version==2.8.5
singledispatch==3.7.0
six==1.15.0
sqlparse==0.4.2
standardjson==0.3.1
swagger-spec-validator==2.7.3
swagger-ui==0.1.2
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
typing_extensions==4.0.1
ua-parser==0.10.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.7
user-agents==2.2.0
virtualenv==20.10.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==1.0.1
wrapt==1.13.3
xlrd==1.2.0
xmltodict==0.12.0
zeep==4.0.0
zipp==3.6.0
